#75 Refactor the app package split it with some sub packages
