+++ web/app/component/websocket.go
@@ -12,7 +12,7 @@
  * the License.
  *******************************************************************************/
 
-package app
+package component
 
 import (
 	_ "encoding/json"
@@ -28,6 +28,9 @@ type WsClientConn struct {
 }
 
 var wsConn *WsClientConn
+
+//target ProxyCache {token:targetIP}
+var DynamicalProxyCache = make(map[string]string, 10)
 
 var upgrader = websocket.Upgrader{
 	ReadBufferSize:  configs.ReadBufferSize,
