+++ web/app/controller/gateway_controller.go
@@ -1,0 +1,63 @@
+/*******************************************************************************
+ * Copyright © 2017-2018 VMware, Inc. All Rights Reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except
+ * in compliance with the License. You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under the License
+ * is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
+ * or implied. See the License for the specific language governing permissions and limitations under
+ * the License.
+ *******************************************************************************/
+
+package controller
+
+import (
+	"encoding/json"
+	"github.com/edgexfoundry-holding/edgex-ui-go/configs"
+	"github.com/edgexfoundry-holding/edgex-ui-go/web/app/component"
+	"github.com/edgexfoundry-holding/edgex-ui-go/web/app/dao"
+	"github.com/edgexfoundry-holding/edgex-ui-go/web/app/domain"
+	"net/http"
+)
+
+func ProxyConfigGateway(w http.ResponseWriter, r *http.Request) {
+	defer r.Body.Close()
+	m := make(map[string]string)
+	err := json.NewDecoder(r.Body).Decode(&m)
+	if err != nil {
+		http.Error(w, err.Error(), http.StatusServiceUnavailable)
+		return
+	}
+	targetIP := m[configs.HostIPKey]
+	component.DynamicalProxyCache[r.Header.Get(configs.SessionTokenKey)] = targetIP
+}
+
+func SaveGateway(w http.ResponseWriter, r *http.Request) {
+	defer r.Body.Close()
+	var g domain.Gateway
+	err := json.NewDecoder(r.Body).Decode(&g)
+	if err != nil {
+		http.Error(w, err.Error(), http.StatusServiceUnavailable)
+		return
+	}
+
+	dao.GetGatewayDao().Add(g, dao.GetDBClient())
+}
+func FindAllGateway(w http.ResponseWriter, r *http.Request) {
+	defer r.Body.Close()
+	w.Header().Set(configs.ContentTypeKey, configs.JsonContentType)
+
+	gateways, err := dao.GetGatewayDao().GetAll(dao.GetDBClient())
+	if err != nil {
+		json.NewEncoder(w).Encode(err.Error())
+		return
+	}
+
+	json.NewEncoder(w).Encode(&gateways)
+}
+func DeleteGateway(w http.ResponseWriter, r *http.Request) {
+
+}
