+++ web/app/service/addressable_service.go
@@ -1,0 +1,7 @@
+//
+// Copyright (c) 2018 Tencent
+//
+// SPDX-License-Identifier: Apache-2.0
+//
+
+package service
